refactor: refactor devspace list vars
