+++ cmd/list/vars.go
@@ -1,9 +1,8 @@
 package list
 
 import (
+	"encoding/json"
 	"fmt"
-	"strings"
-
 	"github.com/loft-sh/devspace/cmd/flags"
 	"github.com/loft-sh/devspace/pkg/util/factory"
 	"github.com/loft-sh/devspace/pkg/util/log"
@@ -14,6 +13,8 @@ import (
 
 type varsCmd struct {
 	*flags.GlobalFlags
+
+	Output string
 }
 
 func newVarsCmd(f factory.Factory, globalFlags *flags.GlobalFlags) *cobra.Command {
@@ -35,6 +36,7 @@ values
 			return cmd.RunListVars(f, cobraCmd, args)
 		}}
 
+	varsCmd.Flags().StringVarP(&cmd.Output, "output", "o", "", "The output format of the command. Can be either empty, keyvalue or json")
 	return varsCmd
 }
 
@@ -57,33 +59,43 @@ func (cmd *varsCmd) RunListVars(f factory.Factory, cobraCmd *cobra.Command, args
 		return err
 	}
 
-	// Get generated config
-	generatedConfig := config.Generated()
-
-	// Specify the table column names
-	headerColumnNames := []string{
-		"Variable",
-		"Value",
-	}
-
-	varRow := make([][]string, 0, len(generatedConfig.Vars))
-	for name, value := range generatedConfig.Vars {
-		if strings.HasPrefix(name, "DEVSPACE_SPACE_DOMAIN") {
-			continue
+	switch cmd.Output {
+	case "":
+		// Specify the table column names
+		headerColumnNames := []string{
+			"Variable",
+			"Value",
 		}
 
-		varRow = append(varRow, []string{
-			name,
-			fmt.Sprintf("%v", value),
-		})
+		varRow := make([][]string, 0, len(config.Variables()))
+		for name, value := range config.Variables() {
+			varRow = append(varRow, []string{
+				name,
+				fmt.Sprintf("%v", value),
+			})
+		}
+
+		// No variable found
+		if len(varRow) == 0 {
+			logger.Info("No variables found")
+			return nil
+		}
+
+		log.PrintTable(logger, headerColumnNames, varRow)
+	case "keyvalue":
+		for name, value := range config.Variables() {
+			fmt.Printf("%s=%v\n", name, value)
+		}
+	case "json":
+		out, err := json.MarshalIndent(config.Variables(), "", "  ")
+		if err != nil {
+			return err
+		}
+
+		fmt.Print(string(out))
+	default:
+		return errors.Errorf("unsupported value for flag --output: %s", cmd.Output)
 	}
 
-	// No variable found
-	if len(varRow) == 0 {
-		logger.Info("No variables found")
-		return nil
-	}
-
-	log.PrintTable(logger, headerColumnNames, varRow)
 	return nil
 }
