+++ pkg/devspace/config/versions/latest/schema.go
@@ -938,7 +938,7 @@ const (
 // ProfileConfig defines a profile config
 type ProfileConfig struct {
 	Name           string                  `yaml:"name" json:"name"`
-	Description    string                  `yaml:"description" json:"description"`
+	Description    string                  `yaml:"description,omitempty" json:"description,omitempty"`
 	Parent         string                  `yaml:"parent,omitempty" json:"parent,omitempty"`
 	Parents        []*ProfileParent        `yaml:"parents,omitempty" json:"parents,omitempty"`
 	Patches        []*PatchConfig          `yaml:"patches,omitempty" json:"patches,omitempty"`
